Create SP13
Applying a security practice "SP13" from the list of pattern getting from research work to a contract (DEX helper contract) living in Ethereum wild. 
NB: The application of the SP13 can be done by different based on this contract structure.  We applied one of them for code part which might be vunrable.